% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean earthquake data frame}
\usage{
eq_clean_data(earthquakes)
}
\arguments{
\item{earthquakes}{A data frame containing the earthquakes data set.}
}
\value{
A new data frame with cleaned earthquakes data.
}
\description{
A DATE column is created based on YEAR, MONTH, DAY.
LATITUDE and LONGITUDE are converted to numeric.
LOCATION_NAME is cleaned by passing it to the \code{\link{eq_location_clean}} function.
}
