% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create HTML labels for earthquakes}
\usage{
eq_create_label(df)
}
\arguments{
\item{df}{Data frame containing at least LOCATION_NAME, EQ_PRIMARY, and
TOTAL_DEATHS columns.}
}
\value{
Character string with the HTML label.
}
\description{
Creates HTML labels with location, magnitude, and total deaths for
earthquake data.
}
\examples{
\dontrun{
# This can be used with \\code{\\link{eq_map}} as follows:
library(dplyr)

earthquakes \%>\%
  dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
  eq_map(annot_col = "popup_text")
}
}
