% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{eq_map}
\alias{eq_map}
\title{Map earthquakes.}
\usage{
eq_map(df, annot_col)
}
\arguments{
\item{df}{Data frame containing at least LONGITUDE, LATITUDE, and EQ_PRIMARY
columns.}

\item{annot_col}{Character string with the name of the column to use for the
popup windows.}
}
\description{
Produces a map of earthquake epicenters (LATITUDE/LONGITUDE) and annotates
each point with a popup window containing annotation data stored in a column
of the data frame. The radius of each circle is proportional to the
earthquake's magnitude (EQ_PRIMARY).
}
\note{
See \code{\link{eq_create_label}} for another example.
}
\examples{
eq_map(earthquakes, annot_col = "DATE")

}
