% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline_geoms.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Labels for earthquake timeline plot}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings specified by aes.}

\item{data}{The earthquake data to be plotted. See the example for details.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the
result of a call to a position adjustment function.}

\item{na.rm}{If FALSE (the default), removes missing values with a
warning. If TRUE silently removes missing values.}

\item{show.legend}{logical. Should this layer be included in the
legends? NA, the default, includes if any aesthetics are
mapped. FALSE never includes, and TRUE always includes.}

\item{inherit.aes}{If FALSE, overrides the default aesthetics,
rather than combining with them.}
}
\description{
This geom is intended to be used in conjunction with the
\code{\link{geom_timeline}} geom to add a vertical line with a text
annotation (e.g. the location of the earthquake) for each data point on an
earthquake timeline. The x aesthetic specifies the date of the earthquake
and the label aesthetic specifies the label for the annotation. An optional
n_max aesthetic can be used to subset to n_max earthquakes, meaning the n_max
largest (by magnitude) will be labeled.
}
\examples{
# You can plot a pair of earthquake timelines with labels as follows:

library(dplyr)

recent_earthquakes <- clean_earthquakes \%>\%
    filter(COUNTRY == "CHINA" | COUNTRY == "USA", DATE >= ymd('2000-01-01'))

g <- ggplot(recent_earthquakes,
            aes(x = DATE, y = COUNTRY, size = EQ_PRIMARY, color = TOTAL_DEATHS))
g <- g + geom_timeline(alpha = 0.5)
g <- g + geom_timeline_label(aes(label = LOCATION_NAME, n_max = 5))
g <- g + theme_classic()
g <- g + theme(legend.position = "bottom",
               axis.line.y = element_blank(),
               axis.ticks.y = element_blank(),
               axis.title.y = element_blank()
g <- g + guides(color = guide_colorbar(title = "# deaths"),
                size = guide_legend("Richter scale value"))
g

}
