% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_date}
\alias{to_date}
\title{Make a Date from year, month, day}
\usage{
to_date(year, month, day)
}
\arguments{
\item{year}{Numeric year}

\item{month}{Numeric month}

\item{day}{Numeric day}
}
\value{
A Date object
}
\description{
This function makes a Date from the supplied year, month, and day paramters.
NA values for month and day are replaced by 1.
}
\details{
Currently requires year to be >= 0.
}
\examples{
to_date(2017, 4, 17)  # => 2017-04-17
to_date(2017, 4, NA)  # => 2017-04-01
to_date(2017, NA, NA) # => 2017-01-01
}
